refactor: move SearchResults to its own file
